Factor out common utility functions into utils.h
